+++ lib/generators/rails_toastify/install/templates/rails_toastify.rb
@@ -1,3 +1,7 @@
-RailsToastify.setup do |config|
-  config.framework = :tailwind #or :bootstrap
+RailsToastify.setup do |configuration|
+  configuration.position = 'toast-container-bottom-right'
+  configuration.animation = 'zoom'
+  configuration.duration = 5000
+  configuration.theme = 'dark'
+  configuration.type = 'success'
 end